Code cleanup

+++ libpsk/include/psk/decoder.hpp
@@ -136,10 +136,10 @@ private:
     _internal::symbol_decoder m_symb_decoder;
 	mode m_rx_mode { mode::bpsk };
     //FIR1 BITFIR_LENGTH
-    dsp::fir_decimate< std::complex<short>,  short, DEC4_LPFIR_LENGTH, std::complex<long> > m_fir1_dec;
-    dsp::fir_decimate< std::complex<short>,  short, DEC4_LPFIR_LENGTH,  std::complex<long> > m_fir2_dec;
-    dsp::fir_decimate< std::complex<short>,  short, BITFIR_LENGTH, std::complex<long> > m_bit_fir;
-    dsp::fir_decimate< std::complex<short>,  short, BITFIR_LENGTH, std::complex<long> > m_freq_fir;
+    dsp::fir_decimate< std::complex<short>,  short, DEC4_LPFIR_LENGTH, std::complex<long long> > m_fir1_dec;
+    dsp::fir_decimate< std::complex<short>,  short, DEC4_LPFIR_LENGTH,  std::complex<long long> > m_fir2_dec;
+    dsp::fir_decimate< std::complex<short>,  short, BITFIR_LENGTH, std::complex<long long> > m_bit_fir;
+    dsp::fir_decimate< std::complex<short>,  short, BITFIR_LENGTH, std::complex<long long> > m_freq_fir;
 };
 
 /* ------------------------------------------------------------------------- */
